+++ src/GafferCycles/IECoreCyclesPreview/ShaderNetworkAlgo.cpp
@@ -354,17 +354,24 @@ T parameterValue( const IECore::CompoundDataMap &parameters, const IECore::Inter
 	return defaultValue;
 }
 
+static const bool g_useLegacyLights = []() -> bool {
+	const char *c = getenv( "GAFFERCYCLES_USE_LEGACY_LIGHTS" );
+	if( !c )
+	{
+		return false;
+	}
+	return strcmp( c, "0" );
+}();
+
 // Cycles lights just have a single `strength` parameter which
 // we want to present as separate "virtual" parameters for
-// intensity, color, exposure and normalize. We calculate un-normalized
-// lights by multiplying the surface area of the light source.
+// intensity, color, exposure.
 bool contributesToLightStrength( InternedString parameterName )
 {
 	return
 		parameterName == "intensity" ||
 		parameterName == "color" ||
-		parameterName == "exposure" ||
-		parameterName == "normalize"
+		parameterName == "exposure"
 	;
 }
 
@@ -387,49 +394,38 @@ Imath::Color3f constantLightStrength( const IECoreScene::ShaderNetwork *light )
 	// you'd want to texture that.
 	strength *= powf( 2.0f, parameterValue<float>( lightShader->parameters(), "exposure", 0.0f ) );
 
-	// Cycles has normalized lights as a default, we can emulate un-normalized lights
-	// with a bit of surface area size calculation onto the strength parameter.
-	/// \todo I think we should move normalization into Cycles itself -
-	/// https://developer.blender.org/D16838
-	if( !parameterValue<bool>( lightShader->parameters(), "normalize", true ) )
-	{
-		if( lightShader->getName() == "distant_light" )
-		{
-			const float angle = IECore::degreesToRadians( parameterValue<float>( lightShader->parameters(), "angle", 0.0f ) ) / 2.0f;
-			const float radius = tanf( angle );
-			const float area = M_PI_F * radius * radius;
-			if( area > 0.0f )
-			{
-				strength *= area;
-			}
-		}
-		else if( lightShader->getName() == "background_light" )
-		{
-			// Do nothing.
-		}
-		else if(
+	if(
+		!g_useLegacyLights &&
+		(
+			lightShader->getName() == "disk_light" ||
+			lightShader->getName() == "point_light" ||
 			lightShader->getName() == "quad_light" ||
-			lightShader->getName() == "portal"
+			lightShader->getName() == "spot_light"
 		)
-		{
-			const float width = parameterValue( lightShader->parameters(), "width", 1.0f );
-			const float height = parameterValue( lightShader->parameters(), "height", 1.0f );
-			strength *= width * height;
-		}
-		else if( lightShader->getName() == "disk_light" )
-		{
-			const float width = parameterValue( lightShader->parameters(), "width", 2.0f ) * 0.5f;
-			const float height = parameterValue( lightShader->parameters(), "height", 2.0f ) * 0.5f;
-			strength *= M_PI * width * height;
-		}
-		else // Point or spot light.
-		{
-			const float size = parameterValue( lightShader->parameters(), "size", 1.0f ) * 0.5f;
-			strength *= M_PI * size * size * 4.0f;
-		}
+	)
+	{
+		// Blender/hdCycles scale by `pi` to convert intensity to radiant flux.
+		// We do the same here to ensure lights exported from Blender to USD render
+		// with an equivalent strength. This also causes these lights to match
+		// the equivalent Arnold light at the same intensity.
+		//
+		// Though all is not perfect and we still have the following quirks :
+		//
+		// Blender scales `distant_light` strength down by 4 when exporting intensity
+		// to USD and scales intensity up by 4 when importing USD or rendering with hdCycles.
+		// For lack of clarity in the UsdLuxLight specification, this value appears to have
+		// been chosen as it approximately matches Karma. Given the lack of clarity here, we
+		// do not apply the same scaling factor. As a result, distant lights render with 4
+		// times less strength than Blender/hdCycles but match Arnold.
+		//
+		// Cycles and Arnold normalize point (sphere) and spotlights differently,
+		// with Arnold lights rendering 4 times brighter than Cycles at the same intensity,
+		// though they match when not normalized. For lack of a definition of standard
+		// behaviour in UsdLuxLight, we don't attempt to match normalization between the two.
+		strength *= M_PI;
 	}
 
-	return strength;;
+	return strength;
 }
 
 const InternedString g_empty( "" );